Revert "Current progress on GFCR"
