+++ .github/workflows/_cdk.yml
@@ -12,7 +12,6 @@ env:
   REACT_APP_AUTH0_DOMAIN: datamermaid.auth0.com
   REACT_APP_CORAL_ATLAS_APP_ID: v8nkxznp3m
   REACT_APP_MERMAID_WHATS_NEW_LINK: https://datamermaid.org/reef-stories/mermaid-v2-what-s-new
-  REACT_APP_GFCR_TEST_USERS: https://datamermaid.org/reef-stories/mermaid-v2-what-s-new
   # dev specific; set prod versions in step below
   REACT_APP_AUTH0_CLIENT_ID: 4AHcVFcwxHb7p1VFB9sFWG52WL7pdNm5
   REACT_APP_AUTH0_AUDIENCE: https://dev-api.datamermaid.org
