+++ src/App/mermaidData/databaseSwitchboard/DatabaseSwitchboard.js
@@ -10,7 +10,6 @@ import SitesMixin from './SitesMixin'
 import FishNameMixin from './FishNamesMixin'
 import SubmittedRecordsMixin from './SubmittedRecordsMixin'
 import ProjectHealthMixin from './ProjectHealthMixin'
-import GfcrMixin from './GfcrMixin'
 
 class DatabaseSwitchboard extends ProjectHealthMixin(
   BenthicAttributesMixin(
@@ -18,9 +17,7 @@ class DatabaseSwitchboard extends ProjectHealthMixin(
       SubmittedRecordsMixin(
         SitesMixin(
           ProjectsMixin(
-            ChoicesMixin(
-              ManagementRegimesMixin(CollectRecordsMixin(GfcrMixin(DatabaseSwitchboardState))),
-            ),
+            ChoicesMixin(ManagementRegimesMixin(CollectRecordsMixin(DatabaseSwitchboardState))),
           ),
         ),
       ),
