--- src/components/GfcrCallout/index.js
@@ -1,3 +1,0 @@
-import GfcrCallout from './GfcrCallout'
-
-export default GfcrCallout
