+++ src/components/generic/ButtonSecondaryDropdown/ButtonSecondaryDropdown.js
@@ -1,9 +1,15 @@
 import PropTypes from 'prop-types'
 import React from 'react'
-import HideShow from '../HideShow'
+import styled from 'styled-components/macro'
+import theme from '../../../theme'
+import HideShow, { DropdownContainer } from '../HideShow'
 import { ButtonSecondary } from '../buttons'
 import { IconDown } from '../../icons'
-import { StyledDropdownContainer } from './ButtonSecondaryDropdown.styles'
+
+const StyledDropdownContainer = styled(DropdownContainer)`
+  background-color: ${theme.color.white};
+  border: solid 1px ${theme.color.border};
+`
 
 const ButtonSecondaryDropdown = ({ children, label, className }) => {
   return (
