+++ src/components/generic/form.js
@@ -83,9 +83,7 @@ export const RequiredIndicator = styled.span`
     content: ' *';
   }
 `
-export const InputRow = styled.div.attrs({
-  className: 'inputRow', // Used to override styles in GFCR forms
-})`
+export const InputRow = styled.div`
   display: grid;
   grid-template-columns: 23rem 31rem 1fr;
   max-width: ${theme.spacing.maxWidth};
@@ -138,9 +136,7 @@ export const Textarea = styled.textarea`
   resize: vertical;
   ${inputTextareaSelectStyles}
 `
-export const LabelContainer = styled.div.attrs({
-  className: 'labelContainer', // Used to override styles in GFCR forms
-})`
+export const LabelContainer = styled.div`
   display: flex;
   flex-direction: row;
   position: relative;
