+++ src/components/icons.js
@@ -16,9 +16,7 @@ import checkIcon from '@iconify-icons/mdi/check'
 import checkOutline from '@iconify-icons/mdi/check-outline'
 import circleEditOutline from '@iconify-icons/mdi/circle-edit-outline'
 import clipboardCheck from '@iconify-icons/mdi/clipboard-check-outline'
-import clipboardEdit from '@iconify-icons/mdi/clipboard-edit'
 import closeIcon from '@iconify-icons/mdi/close'
-import closeIconCircle from '@iconify-icons/mdi/close-circle'
 import contentCopy from '@iconify-icons/mdi/content-copy'
 import contentSave from '@iconify-icons/mdi/content-save'
 import download from '@iconify-icons/mdi/download'
@@ -77,7 +75,6 @@ export const IconCheck = (props) => <InlineIcon icon={checkIcon} {...props} />
 export const IconCheckOutline = (props) => <InlineIcon icon={checkOutline} {...props} />
 export const IconCheckAll = (props) => <InlineIcon icon={checkAllIcon} {...props} />
 export const IconClose = (props) => <InlineIcon icon={closeIcon} {...props} />
-export const IconCloseCircle = (props) => <InlineIcon icon={closeIconCircle} {...props} />
 export const IconCollect = (props) => <InlineIcon icon={circleEditOutline} {...props} />
 export const IconCopy = (props) => <InlineIcon icon={contentCopy} {...props} />
 export const IconData = (props) => <InlineIcon icon={checkCircleOutline} {...props} />
@@ -86,7 +83,6 @@ export const IconDownload = (props) => <InlineIcon icon={download} {...props} />
 export const IconExternalLink = (props) => <InlineIcon icon={launch} {...props} />
 export const IconFilter = (props) => <InlineIcon icon={filterOutline} {...props} />
 export const IconFish = (props) => <InlineIcon icon={fish} {...props} />
-export const IconGfcr = (props) => <InlineIcon icon={clipboardEdit} {...props} />
 export const IconGlobe = (props) => <InlineIcon icon={globe} {...props} />
 export const IconGraph = (props) => <InlineIcon icon={chartBar} {...props} />
 export const IconHeart = (props) => <InlineIcon icon={heartOutline} {...props} />
