+++ src/components/pages/Site/Site.js
@@ -48,7 +48,6 @@ import InputSelectWithLabelAndValidation from '../../mermaidInputs/InputSelectWi
 import { DeleteRecordButtonCautionWrapper } from '../collectRecordFormPages/CollectingFormPage.Styles'
 import { IconSwap } from '../../icons'
 import { InputButton } from '../../generic/buttons'
-import { enforceNumberInput } from '../../../library/enforceNumberInput'
 
 const ReadOnlySiteContent = ({
   site,
@@ -92,6 +91,28 @@ const SwapButton = ({ isDisabled, handleSwapClick, swapLabel }) => {
       <span>{swapLabel}</span>
     </InputButton>
   )
+}
+
+const enforceNumberInput = (event) => {
+  // Allow only numbers, special key presses, and copy paste shortcuts.
+  const specialActionAndCharacterKeys = [
+    'ArrowUp',
+    'ArrowDown',
+    'ArrowLeft',
+    'ArrowRight',
+    'Backspace',
+    'Delete',
+    '-',
+    '.',
+    'Tab',
+  ]
+
+  const isModifiersKeyPressed = event.metaKey || event.ctrlKey || event.shiftKey
+  const isMovingAndSpecialCharactersKeyPressed = specialActionAndCharacterKeys.includes(event.key)
+  const isNumbersKeyPressed =
+    (event.keyCode >= 48 && event.keyCode <= 58) || (event.keyCode >= 96 && event.keyCode <= 105)
+
+  return !(isModifiersKeyPressed || isMovingAndSpecialCharactersKeyPressed || isNumbersKeyPressed)
 }
 
 const SiteForm = ({
