--- src/components/pages/gfcrPages/Gfcr/Gfcr.stories.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-import Gfcr from '.'
-
-export default {
-  title: 'GFCR',
-  component: Gfcr,
-}
-export const basic = () => <Gfcr />
