--- src/components/pages/gfcrPages/Gfcr/index.js
@@ -1,3 +1,0 @@
-import Gfcr from './Gfcr'
-
-export default Gfcr
