--- src/components/pages/gfcrPages/GfcrIndicatorSet/index.js
@@ -1,3 +1,0 @@
-import GfcrIndicatorSet from './GfcrIndicatorSet'
-
-export default GfcrIndicatorSet
