--- src/components/pages/gfcrPages/GfcrIndicatorSetForm/index.js
@@ -1,3 +1,0 @@
-import GfcrIndicatorSetForm from './GfcrIndicatorSetForm'
-
-export default GfcrIndicatorSetForm
