--- src/components/pages/gfcrPages/GfcrIndicatorSetForm/subForms/index.js
@@ -1,6 +1,0 @@
-import IndicatorSetForm from './IndicatorSetForm'
-import F1Form from './F1Form'
-import F2Form from './F2Form'
-import F3Form from './F3Form'
-
-export { IndicatorSetForm, F1Form, F2Form, F3Form }
