--- src/components/pages/gfcrPages/GfcrIndicatorSetNav/index.js
@@ -1,3 +1,0 @@
-import GfcrIndicatorSetNav from './GfcrIndicatorSetNav'
-
-export default GfcrIndicatorSetNav
