+++ src/language.js
@@ -72,8 +72,6 @@ const error = {
     `The ${mermaidDataTypeLabel} failed to save both on your computer and online.`,
   getDeleteOfflineErrorTitle: (mermaidDataTypeLabel) =>
     `The ${mermaidDataTypeLabel} has failed to delete from your computer or online.`,
-  gfcrIndicatorSetsUnavailable: 'GFCR indicator sets are currently unavailable.',
-  gfcrIndicatorSetSave: 'Indicator set has not been saved.',
   idNotFoundUserAction: "Please check the URL in your browser's address bar.",
   invalidEmailAdd: 'Invalid email address.',
   managementRegimeRecordsUnavailable: 'Management Regime records data are currently unavailable.',
@@ -82,8 +80,6 @@ const error = {
   notificationNotDeleted: 'Notification could not be removed.',
   notificationsNotDeleted: 'Notifications could not be removed',
   projectDelete: 'This project has not been deleted.',
-  projectAddGfcr: 'Could not add GFCR indicators to the project.',
-  projectRemoveGfcr: 'Could not remove GFCR indicators from the project.',
   projectSave: 'The project has not been saved.',
   projectsUnavailable: 'Project data are currently unavailable.',
   projectWithSameName: 'A project with the same name already exists.',
@@ -119,7 +115,6 @@ const error = {
   pageNotFound: 'This page cannot be found.',
   pageNotFoundRecovery: 'Make sure the URL is correct.',
   pageReadOnly: 'You cannot access this page because you are a read-only member of this project.',
-  pageAdminOnly: 'You cannot access this page because you are not an admin for this project.',
   idNotFound: 'This item cannot be found.',
   idNotFoundRecovery:
     'It might have been deleted, you do not have permission to view it, or the URL might be wrong.',
@@ -150,13 +145,10 @@ const success = {
     `${projectName} has been removed from being offline ready`,
   getUserRoleChangeSuccessMessage: ({ userName, role }) =>
     `${userName}'s role is now set to ${role}.`,
-  gfcrIndicatorSetSave: 'Indicator set saved.',
   newUserAdd: 'New user added.',
   newPendingUserAdd: 'Sign-up email sent. New user added as Pending User.',
   userRemoved: 'User removed',
   projectSave: 'Project saved',
-  projectAddGfcr: 'Added GFCR indicators to project',
-  projectRemoveGfcr: 'Removed GFCR indicators from project',
   projectCopied: 'Project copied',
   projectCreated: 'Project created',
   projectDeleted: 'Project deleted',
@@ -358,13 +350,6 @@ const pages = {
     organizations: 'Organizations',
     notes: 'Notes',
     noOrganization: 'This Project has no organizations.',
-    gfcrCalloutHeading: 'Global Fund for Coral Reefs (GFCR)',
-    gfcrRemoveParagraph:
-      'Removing GFCR indicators from this project will not delete them, but just hide them.',
-    gfcrAddParagraph:
-      'GFCR is a global partnership that aims to mobilize resources to support coral reef conservation and restoration projects around the world. ',
-    gfcrRemoveButton: 'Remove GFCR indicators from this project',
-    gfcrAddButton: 'Add GFCR indicators to this project',
   },
   dataSharing: {
     introductionParagraph: `Given the urgent need for global coral reef conservation, MERMAID is committed to working collectively as a community and using the power of data to help make faster, better decisions. Coral reef monitoring data are collected with the intent of advancing coral reef science and improving management. We recognize the large effort to collect data and your sense of ownership. While not required, we hope you choose to make your data available to fuel new discoveries and inform conservation solutions.`,
@@ -463,50 +448,6 @@ const pages = {
     title: 'Copy Management Regimes',
     filterToolbarText: 'Filter management regimes by name, project, or year',
     copyButtonText: 'Copy selected MRs to project',
-  },
-  gfcrTable: {
-    filterToolbarText: 'Filter this table by indicator set, and other things?',
-    title: 'GFCR',
-    noDataMainText: 'No indicator sets yet.',
-    noDataSubText: "Select 'Create new' to add an indicator set to this project.",
-  },
-  gfcrIndicatorSet: {
-    title: 'Indicator Set',
-    f1_1: 'Coral reef area of coral reefs in GFCR programming',
-    f2_1a: 'Coral reef area of MPAs and OECMs (as aligned to GBF Target 3)',
-    f2_1b: 'Total area of MPAs and OECMs (as aligned to GBF Target 3)',
-    f2_2a: 'Coral reef area of locally managed areas / co-managed areas',
-    f2_2b: 'Total area of locally managed areas / co-managed areas',
-    f2_3a: 'Coral reef area of fisheries management',
-    f2_3b: 'Total area of fisheries management',
-    f2_4: 'Area with pollution mitigation',
-    f2_opt1:
-      'Area of non-coral reef ecosystems, e.g. mangroves, seagrass or other associated ecosystems',
-    f3_1: 'Area of effective coral reef restoration',
-    f3_2: 'Number of in situ coral restoration projects',
-    f3_3: 'Number of coral restoration plans, technologies, strategies or guidelines developed',
-    f3_4: 'Number of coral restoration trainings',
-    f3_5: 'Number of people engaged in coral restoration',
-    f3_5_total: 'Total',
-    f3_5_ofTotalHowMany: 'Of the total, how many of the following',
-    f3_5a: 'Men',
-    f3_5b: 'Women',
-    f3_5c: 'Youth',
-    f3_5d: 'Indigenous',
-    f3_6: 'Number of response plans (incl. financial mechanisms, eg., insurance) in place to support coral restoration after severe shocks (e.g,. storms, bleaching)',
-  },
-  gfcrIndicatorSetNav: {
-    indicatorSetHeading: 'Indicator Set',
-    f1: 'Coral reef extent of GFCR project',
-    f2: 'Area of coral reefs under conservation and sustainable management',
-    f3: 'Area of coral reefs under effective coral restoration',
-    f4: 'Change in coral reef health',
-    f5: 'Number of communities engaged in meaningful participation, co-development and capacity strengthening',
-    f6: 'Number of people supported through livelihoods, direct jobs, income, and nutrition',
-    f7: 'Number of people supported to better adapt, respond and recover to the effects of climate change and major external shocks as a result of GFCR',
-    financialData: 'Financial Data (F8, F9, F10)',
-    investmentSources: 'Investment Sources',
-    revenueStreams: 'Revenue Streams',
   },
   goToDashboard: 'View on Dashboard',
 }
